Add changes from http://github.com/nickgammon/mushclient/commit/9151f01
1. During plugin callbacks attempts to output Note, Tell or similar (eg. ColourNote, print in Lua) are deferred until the end of the next line of input from the MUD.

This is to prevent things like MXP callbacks, or telnet negotiation callbacks (like GMCP), which output debugging or informational messages from inadvertently turning the current line in the output buffer from "MUD output" status to "Note status".

A side-effect of such status conversion is that the next (or current) line would skip trigger-matching.